+++ src/components/Medications/data.js
@@ -1,0 +1,73 @@
+const data = [
+  {
+  name: 'Allesse', 
+  type: 'Extended Cycle', 
+  efforts: 'Medium', 
+  symptoms:  
+  [ 'nausea', 
+    'changes to eyesight for those using contact lenses', 
+    'mood changes', 
+    'mood changes', 
+    'changes to eyesight for those using contact lenses', 
+    'breast tenderness', 
+    'nausea' ]
+  }, 
+  {
+  name: 'Aranelle', 
+  type: 'Combo', 
+  efforts: 'Low', 
+  symptoms:  
+  [ 'nausea', 
+    'nausea', 
+    'Change in corneal curvature (steepening)', 
+    'intermenstrual spotting', 
+    'changes to eyesight for those using contact lenses', 
+    'Melasma (dark pigmentation on the face) which may persist', 
+    'Reduced tolerance to carbohydrates (insulin resistance)' ]
+  }, 
+  {
+    name: 'Aranelle', 
+    type: 'Combo', 
+    efforts: 'High', 
+    symptoms:  [
+      'intermenstrual spotting', 
+      'headaches and migraine', 
+      'decreased libido', 
+      'Reduced tolerance to carbohydrates (insulin resistance)', 
+      'Change in corneal curvature (steepening)', 
+      'intermenstrual spotting', 
+      'missed periods'
+    ]
+  }, 
+  {
+    name: 'Aranelle', 
+    type: 'Oral', 
+    efforts: 'High', 
+    symptoms: [
+      'missed periods', 
+      'weight gain', 
+      'decreased libido', 
+      'headaches and migraine', 
+      'changes to eyesight for those using contact lenses', 
+      'Melasma (dark pigmentation on the face) which may persist', 
+      'Vaginal candidiasis'
+    ]
+  }, 
+  {
+    name: 'Caziant', 
+    type: 'Oral', 
+    efforts: 'Medium', 
+    symptoms: [
+      'changes to eyesight for those using contact lenses', 
+      'Mental depression', 
+      'Reduced tolerance to carbohydrates (insulin resistance)', 
+      'headaches and migraine', 
+      'Vaginal candidiasis', 
+      'weight gain', 
+      'Abdominal cramps and bloating'
+    ]
+  }
+];
+
+export default data;
+
